validation logic extraction to seperate file
