+++ .azure/build.yaml
@@ -19,7 +19,7 @@ steps:
       pip install pyyaml
     displayName: 'Install dependencies'
 
-  - script: python .azure/process_metadata.py 
+  - script: python .azure/scripts/process_metadata.py 
     displayName: 'Run metadata extraction script'
 
   - task: AzureCLI@2
