+++ .azure/pr-check.yaml
@@ -20,7 +20,7 @@ steps:
     displayName: "Install dependencies"
 
   - script: |
-      python .azure/validate_samples.py
+      python .azure/scripts/validate_samples.py
     displayName: "Metadata Validation"
 
   - script: echo 'Pull Request Validation Completed'
