+++ .azure/scripts/process_metadata.py
@@ -2,22 +2,12 @@ import os
 import yaml
 import json
 import shutil
+import metadata_validator
 
 REPO_BASE_DIR = os.environ['BUILD_SOURCESDIRECTORY']
 BUILD_STAGING_DIRECTORY = os.environ['BUILD_STAGINGDIRECTORY']
 BASE_URL_FOR_THUMBNAILS = 'https://choreo-shared-choreo-samples-cdne.azureedge.net'
-WSO2_SAMPLES_REPO_URL = 'https://github.com/wso2/choreo-samples/'
 
-VALID_COMPONENT_TYPES = [
-    "service", "webhook", "manual-task", "scheduled-task", 
-    "event-triggered", "event-handler", "test-runner", "many",
-    "web-application"
-]
-
-VALID_BUILD_PACKS = [
-    "ballerina", "wso2-mi", "go", "java", "php", "python", "nodejs", "ruby", 
-    "many", "postman", "react", "docker"
-]
 
 def collect_metadata_and_thumbnails():
     collected_data = []
@@ -35,19 +25,18 @@ def collect_metadata_and_thumbnails():
 
                 component_path = data.get('componentPath')
                 # Check if the componentPath exists
-                if not os.path.exists(os.path.join(REPO_BASE_DIR, component_path.lstrip('/'))):
-                    if WSO2_SAMPLES_REPO_URL == data.get('repositoryUrl'):
-                        print(f"Warning: Component path '{component_path}' does not exist. This will be excluded from index.json.")
-                        continue
+                if not metadata_validator.validate_component_path(component_path, data.get('repositoryUrl')):
+                    print(f"Warning: Component path '{component_path}' does not exist. This will be excluded from index.json.")
+                    continue
 
                 component_type = data.get('componentType', '')
                 build_pack = data.get('buildPack', '')
 
-                if component_type not in VALID_COMPONENT_TYPES:
+                if not metadata_validator.validate_component_type(component_type):
                     print(f"Warning: '{component_type}' is not a valid componentType for directory: {meta_file}. Excluding from index.json.")
                     continue
 
-                if build_pack not in VALID_BUILD_PACKS:
+                if not metadata_validator.validate_build_pack(build_pack):
                     print(f"Warning: '{build_pack}' is not a valid buildPreset for directory: {meta_file}. Excluding from index.json.")
                     continue
 
@@ -59,7 +48,7 @@ def collect_metadata_and_thumbnails():
             thumbnail_src = os.path.join(samples_dir, data.get('thumbnailPath').lstrip('/'))
             thumbnail_dest_folder = os.path.join(BUILD_STAGING_DIRECTORY, 'icons')
             os.makedirs(thumbnail_dest_folder, exist_ok=True)
-            if os.path.exists(thumbnail_src):
+            if metadata_validator.validate_thumbnail(thumbnail_src):
                 shutil.copy(thumbnail_src, thumbnail_dest_folder)
             else:
                 print(f"Thumbnail not found for {thumbnail_src}")
@@ -71,12 +60,9 @@ def collect_metadata_and_thumbnails():
             collected_data.append(data)
 
     # Check if there are any directories without corresponding metadata files
-    for dir_name in os.listdir(REPO_BASE_DIR):
-        dir_path = os.path.join(REPO_BASE_DIR, dir_name)
-        if os.path.isfile(dir_path) or dir_name.startswith('.'):
-            continue
-        if dir_name not in samples_dirnames_set:
-            print(f"Warning: Directory '{dir_name}' does not have a corresponding metadata file. This will be excluded form index.json.")
+    is_valid, dir_name = metadata_validator.validate_directories_for_metafiles(samples_dirnames_set)
+    if not is_valid:
+        print(f"Warning: Directory '{dir_name}' does not have a corresponding metadata file. This will be excluded form index.json.")
 
     return collected_data
